Restored lost code from last deleted repo

+++ WebShop/Models/Managers/ProductManager.cs
@@ -1,4 +1,5 @@
-﻿using WebShop.Data;
+﻿using Microsoft.EntityFrameworkCore;
+using WebShop.Data;
 using WebShop.Models.Interfaces;
 
 namespace WebShop.Models.Managers
@@ -6,13 +7,18 @@ namespace WebShop.Models.Managers
     public class ProductManager : IProductManager
     {
 
-        public async Task<Product> GetProduct(int id)
+        public async Task<Product?> GetProduct(int id)
         {
 
             try
             {
                 using var db = new AppDbContext();
-                return db.Products.FirstOrDefault(p => p.Id == id);
+                var p = await db.Products.FirstOrDefaultAsync(p => p.Id == id);
+                if (p != null)
+                {
+                    return p;
+                }
+                return null;
 
             }
             catch (Exception e)
@@ -23,11 +29,15 @@ namespace WebShop.Models.Managers
         }
         public async Task AddProduct(Product product)
         {
+            if (product == null)
+            {
+                return;
+            }
             try
             {
 
                 using var db = new AppDbContext();
-                db.Products.Add(product);
+                await db.Products.AddAsync(product);
             }
             catch (Exception e)
             {
@@ -41,7 +51,11 @@ namespace WebShop.Models.Managers
             {
 
                 using var db = new AppDbContext();
-                var oldProductFromList = db.Products.FirstOrDefault(p => p.Id == oldProduct.Id);
+                var oldProductFromList = await db.Products.FirstOrDefaultAsync(p => p.Id == oldProduct.Id);
+                if (oldProductFromList == null)
+                {
+                    return;
+                }
                 oldProductFromList = newProduct;
                 await db.SaveChangesAsync();
             }
@@ -58,9 +72,32 @@ namespace WebShop.Models.Managers
             {
 
                 using var db = new AppDbContext();
-                var p = db.Products.FirstOrDefault(p => p.Id == id);
+                var p = await db.Products.FirstOrDefaultAsync(p => p.Id == id);
+                if (p == null)
+                {
+                    return;
+                }
                 db.Products.Remove(p);
                 await db.SaveChangesAsync();
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e.Message);
+                throw;
+            }
+        }
+
+        public async Task<List<Product>?> ProductsToList()
+        {
+            try
+            {
+                using var db = new AppDbContext();
+                var p = await db.Products.ToListAsync();
+                if (p != null)
+                {
+                    return p;
+                }
+                return null;
             }
             catch (Exception e)
             {
@@ -75,7 +112,7 @@ namespace WebShop.Models.Managers
             {
 
                 using var db = new AppDbContext();
-                db.Categories.Add(new Category { Name = category });
+                await db.Categories.AddAsync(new Category { Name = category });
                 await db.SaveChangesAsync();
             }
             catch (Exception e)
@@ -91,8 +128,31 @@ namespace WebShop.Models.Managers
             {
 
                 using var db = new AppDbContext();
-                var c = db.Categories.ToList();
+                var c = await db.Categories.ToListAsync();
                 return c;
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e.Message);
+                throw;
+            }
+        }
+
+        public async Task<Category?> SearchCategory(string category)
+        {
+            if (category == null)
+            {
+                return null;
+            }
+            try
+            {
+                using var db = new AppDbContext();
+                var c = await db.Categories.FirstOrDefaultAsync(c => c.Name == category);
+                if (c != null)
+                {
+                    return c;
+                }
+                return null;
             }
             catch (Exception e)
             {
